a lot of new

+++ PhotoStudio/StaticData.cs
@@ -10,9 +10,8 @@ namespace PhotoStudio
     {
         public static List<string> PackageList = new List<string>();
         public static List<string> EmployeeList = new List<string>();
-        public static Dictionary<string, int> Package = new Dictionary<string, int>();
+        public static Dictionary<string, float> Package = new Dictionary<string, float>();
         public static Dictionary<string, int> Employee = new Dictionary<string, int>();
         public static Dictionary<int, float> ClassEmployee = new Dictionary<int, float>();
-
     }
 }
